feat(new-trace-view): Moved traceheader component to it's own file.
